Refactor : Basic Code Refactoring #1
1. RefreshToken Entity삭제
