+++ src/main/java/com/solver/solver_be/domain/user/service/GuestService.java
@@ -9,7 +9,7 @@ import com.solver.solver_be.domain.user.repository.GuestRepository;
 import com.solver.solver_be.global.exception.exceptionType.UserException;
 import com.solver.solver_be.global.response.GlobalResponseDto;
 import com.solver.solver_be.global.security.jwt.JwtUtil;
-import com.solver.solver_be.global.security.refreshtoken.TokenDto;
+import com.solver.solver_be.global.security.jwt.dto.TokenDto;
 import com.solver.solver_be.global.type.ErrorType;
 import com.solver.solver_be.global.type.SuccessType;
 import com.solver.solver_be.global.type.UserRoleEnum;
