--- src/main/java/com/solver/solver_be/global/security/refreshtoken/RefreshToken.java
@@ -1,26 +1,0 @@
-package com.solver.solver_be.global.security.refreshtoken;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.validation.constraints.NotBlank;
-
-@Getter
-@Entity
-@NoArgsConstructor
-public class RefreshToken {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NotBlank
-    private String refreshToken;
-
-    @NotBlank
-    private String userEmail;
-}
