--- src/main/java/com/solver/solver_be/global/security/refreshtoken/RefreshTokenRepository.java
@@ -1,6 +1,0 @@
-package com.solver.solver_be.global.security.refreshtoken;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
-}
